Trait rivulet::splittable::SplittableImpl

pub unsafe trait SplittableImpl {
    type Item;
    type Error: Debug;
    unsafe fn set_reader_waker(
        &mut self, 
        waker: impl Fn() + Send + Sync + 'static
    );

    unsafe fn set_head(&mut self, index: u64);

    unsafe fn compare_set_head(&self, index: u64);

    fn poll_available(
        self: Pin<&Self>, 
        cx: &mut Context<'_>, 
        register_wakeup: impl FnOnce(&Waker), 
        index: u64, 
        len: usize
    ) -> Poll<Result<usize, Self::Error>>;

    unsafe fn view(&self, index: u64, len: usize) -> &[Self::Item];
}

The implementation behind Splittable.

Unless you are manually implementing a source, you should use Splittable directly.

Associated Types

type Item

The streamed type.

type Error: Debug

The error produced by poll_available.

Required methods

unsafe fn set_reader_waker(&mut self, waker: impl Fn() + Send + Sync + 'static)

Set the reader waking function.

This should only be called once, after creation.

Safety

Only set the waker if you have unique ownership of this.

unsafe fn set_head(&mut self, index: u64)

Set the earliest position retained in the stream.

Panics

May panic if the provided an index earlier than the current head.

Safety

Only set the head if you have unique ownership of this and will not attempt to read any data earlier than index.

unsafe fn compare_set_head(&self, index: u64)

Set the earliest position retained in the stream.

This function, unlike set_head, is synchronized between threads. If the provided head is less than the current head, the current head remains unchanged.

Safety

See set_head.

fn poll_available(
self: Pin<&Self>,
cx: &mut Context<'_>,
register_wakeup: impl FnOnce(&Waker),
index: u64,
len: usize
) -> Poll<Result<usize, Self::Error>>

Suspends the current task until len samples starting at index are available, returning the available length.

If the stream is closed, the returned available length may be less than len.

unsafe fn view(&self, index: u64, len: usize) -> &[Self::Item]

Obtain a view into the stream.

Safety

The parameters must be within an available window as returned by poll_available.