Trait rivulet::splittable::SplittableImplMut

pub unsafe trait SplittableImplMut: SplittableImpl {
    unsafe fn view_mut(&self, index: u64, len: usize) -> &mut [Self::Item];
}

The implementation behind SplittableMut.

Unless you are manually implementing a view, you should use SplittableMut directly.

Required methods

unsafe fn view_mut(&self, index: u64, len: usize) -> &mut [Self::Item]

Obtain a mutable view into the stream.

Safety

The parameters must be within an available window as returned by poll_available.

The view must not overlap with any other view of this stream.

Note that this function produces a mutable reference from a regular reference. Implementations of this function must take care to correctly implement interior mutability (such as using UnsafeCell).