Trait rivulet::splittable::SplittableView

pub trait SplittableView: SplittableViewImpl {
    // Provided methods
    fn into_view(self) -> View<Self> { ... }
    fn into_cloneable_view(self) -> Cloneable<Self> { ... }
    fn sequence(self) -> (First<Self>, Second<Self>) { ... }
}

A view that can be split for use with multiple readers.

Provided Methods

fn into_view(self) -> View<Self>

Create a view for a single reader.

fn into_cloneable_view(self) -> Cloneable<Self>

Create a view that implements Clone.

fn sequence(self) -> (First<Self>, Second<Self>)

Split this view into two sequential views, such that data released by First becomes accessible to Second.