Trait rivulet::Sink

pub trait Sink {
    type Item;
    fn sink(&mut self) -> &mut [Self::Item];

    fn poll_reserve(
        self: Pin<&mut Self>, 
        cx: &mut Context<'_>, 
        count: usize
    ) -> Poll<Result<(), Error>>;

    fn poll_commit(
        self: Pin<&mut Self>, 
        cx: &mut Context<'_>, 
        count: usize
    ) -> Poll<Result<(), Error>>;

    fn reserve(&mut self, count: usize) -> Reserve<'_, Self>
    where
        Self: Sized + Unpin,
    { ... }

    fn commit(&mut self, count: usize) -> Commit<'_, Self>
    where
        Self: Sized + Unpin,
    { ... }

    fn blocking_reserve(&mut self, count: usize) -> Result<(), Error>
    where
        Self: Sized + Unpin,
    { ... }

    fn blocking_commit(&mut self, count: usize) -> Result<(), Error>
    where
        Self: Sized + Unpin,
    { ... }
}

Interface for asynchronous contiguous-memory sinks.

Implementors of Sink may be called "writers".

Associated Types

type Item

The type to be written.

Loading content...
Required methods

fn sink(&mut self) -> &mut [Self::Item]

The mutable buffer for writing data.

This buffer is obtained by successfully polling poll_reserve and committed by successfully polling poll_commit.

fn poll_reserve(
self: Pin<&mut Self>,
cx: &mut Context<'_>,
count: usize
) -> Poll<Result<(), Error>>

Attempt to reserve at least count elements in the writable buffer.

fn poll_commit(
self: Pin<&mut Self>,
cx: &mut Context<'_>,
count: usize
) -> Poll<Result<(), Error>>

Attempt to commit the first count elements in the writable buffer to the stream.

Loading content...
Provided methods

fn reserve(&mut self, count: usize) -> Reserve<'_, Self> where
Self: Sized + Unpin,

Create a future that reserves at least count elements in the writable buffer.

See poll_reserve.

fn commit(&mut self, count: usize) -> Commit<'_, Self> where
Self: Sized + Unpin,

Create a future that commits the first count elements in the writable buffer to the stream.

See poll_commit.

fn blocking_reserve(&mut self, count: usize) -> Result<(), Error> where
Self: Sized + Unpin,

Reserve at least count elements in the writable buffer, blocking the current thread.

See poll_reserve.

fn blocking_commit(&mut self, count: usize) -> Result<(), Error> where
Self: Sized + Unpin,

Commit the first count elements in the writable buffer to the stream, blocking the current thread.

See poll_commit.

Loading content...