Trait rivulet::Source

pub trait Source {
    type Item;
    fn source(&self) -> &[Self::Item];

    fn poll_request(
        self: Pin<&mut Self>, 
        cx: &mut Context<'_>, 
        count: usize
    ) -> Poll<Result<(), Error>>;

    fn poll_consume(
        self: Pin<&mut Self>, 
        cx: &mut Context<'_>, 
        count: usize
    ) -> Poll<Result<(), Error>>;

    fn request(&mut self, count: usize) -> Request<'_, Self>
    where
        Self: Sized + Unpin,
    { ... }

    fn consume(&mut self, count: usize) -> Consume<'_, Self>
    where
        Self: Sized + Unpin,
    { ... }

    fn blocking_request(&mut self, count: usize) -> Result<(), Error>
    where
        Self: Sized + Unpin,
    { ... }

    fn blocking_consume(&mut self, count: usize) -> Result<(), Error>
    where
        Self: Sized + Unpin,
    { ... }
}

Interface for asynchronous contiguous-memory sources.

Implementors of Source may be called "readers".

Associated Types

type Item

The type to be read.

Loading content...
Required methods

fn source(&self) -> &[Self::Item]

The buffer for reading data.

This buffer is obtained by successfully polling poll_request and advanced by successfully polling poll_consume.

fn poll_request(
self: Pin<&mut Self>,
cx: &mut Context<'_>,
count: usize
) -> Poll<Result<(), Error>>

Attempt to read at least count elements into the buffer.

fn poll_consume(
self: Pin<&mut Self>,
cx: &mut Context<'_>,
count: usize
) -> Poll<Result<(), Error>>

Attempt to advance past the first count elements in the buffer.

Loading content...
Provided methods

fn request(&mut self, count: usize) -> Request<'_, Self> where
Self: Sized + Unpin,

Create a future that reads at least count elements into the buffer.

See poll_request.

fn consume(&mut self, count: usize) -> Consume<'_, Self> where
Self: Sized + Unpin,

Create a future that advances past the first count elements in the buffer.

See poll_consume.

fn blocking_request(&mut self, count: usize) -> Result<(), Error> where
Self: Sized + Unpin,

Reads at least count elements into the buffer, blocking the current thread.

See poll_request.

fn blocking_consume(&mut self, count: usize) -> Result<(), Error> where
Self: Sized + Unpin,

Advances past the first count elements in the buffer, blocking the current thread.

See poll_consume.

Loading content...