Trait rivulet::StableView

pub unsafe trait StableView: View {
    fn stable_view(&self, offset: usize, len: usize) -> *const Self::Item;
}

Obtain views with stable memory locations.

Safety

The lifetime of the pointers returned by this trait must be at least as long as the subview remains in the current view.

Required methods

fn stable_view(&self, offset: usize, len: usize) -> *const Self::Item

Obtain a stable subview of the current view, starting at offset elements from the start of the current view and spanning len elements.

The subview has the following qualities:

It may not be written to.

Adjacent or overlapping subviews are not guaranteed to be adjacent or overlapping in memory.

Dereferencing the pointer after the view is advanced past the start of the subview is undefined behavior.

Panics

Panics if the subview would extend beyond the current view.

Loading content...