Trait rivulet::View

pub trait View {
    type Item;
    fn view(&self) -> &[Self::Item];

    fn poll_grant(
        self: Pin<&mut Self>, 
        cx: &mut Context<'_>, 
        count: usize
    ) -> Poll<Result<(), Error>>;

    fn poll_release(
        self: Pin<&mut Self>, 
        cx: &mut Context<'_>, 
        count: usize
    ) -> Poll<Result<(), Error>>;

    fn grant(&mut self, count: usize) -> Grant<'_, Self>
    where
        Self: Sized + Unpin,
    { ... }

    fn release(&mut self, count: usize) -> Release<'_, Self>
    where
        Self: Sized + Unpin,
    { ... }

    fn blocking_grant(&mut self, count: usize) -> Result<(), Error>
    where
        Self: Sized + Unpin,
    { ... }

    fn blocking_release(&mut self, count: usize) -> Result<(), Error>
    where
        Self: Sized + Unpin,
    { ... }
}

Obtain views into asynchronous contiguous-memory streams.

Associated Types

type Item

The streamed type.

Loading content...
Required methods

fn view(&self) -> &[Self::Item]

Obtain the current view of the stream.

This view is obtained by successfully polling poll_grant and advanced by successfully polling poll_release.

fn poll_grant(
self: Pin<&mut Self>,
cx: &mut Context<'_>,
count: usize
) -> Poll<Result<(), Error>>

Attempt to obtain a view of at least count elements.

fn poll_release(
self: Pin<&mut Self>,
cx: &mut Context<'_>,
count: usize
) -> Poll<Result<(), Error>>

Attempt to advance past the first count elements in the current view.

Loading content...
Provided methods

fn grant(&mut self, count: usize) -> Grant<'_, Self> where
Self: Sized + Unpin,

Create a future that obtains a view of at least count elements.

See poll_grant.

fn release(&mut self, count: usize) -> Release<'_, Self> where
Self: Sized + Unpin,

Create a future that advances past the first count elements in the current view.

See poll_release.

fn blocking_grant(&mut self, count: usize) -> Result<(), Error> where
Self: Sized + Unpin,

Obtains a view of at least count elements, blocking the current thread.

See poll_grant.

fn blocking_release(&mut self, count: usize) -> Result<(), Error> where
Self: Sized + Unpin,

Advances past the first count elements in the current view, blocking the current thread.

See poll_release.

Loading content...