Crate rivulet

Rivulet provides tools for creating and processing asynchronous streams of contiguous data.

Modules

circular_bufferstd

Asynchronous circular buffers.

error

Errors produced by streams.

iostd

Utilities for working with std::io.

lazy

Lazy-initialized streams.

slice

Utilities for working with slices.

splittable

Streams that can be split into multiple sources.

Structs

Grant

Future produced by View::grant.

MapError

An error-mapped view produced by View::map_error.

Traits

Sink

A marker trait that indicates this view is a sink of data.

Source

A marker trait that indicates this view is a source of data.

View

Obtain views into asynchronous contiguous-memory streams.

ViewMut

Obtain mutable views into asynchronous contiguous-memory mutable streams.