Module rivulet::splittable

Streams that can be split into multiple views.

Structs

Cloneable

A view returned by SplittableView::into_cloneable_view.

First

The first SplittableView produced by sequence.

Second

The second SplittableView produced by sequence.

View

A view returned by SplittableView::into_view.

Traits

SplittableView

A view that can be split for use with multiple readers.

SplittableViewImpl

The implementation behind SplittableView.

SplittableViewImplMut

The implementation behind SplittableViewMut.

SplittableViewMut

A mutable view that can be split for use with multiple readers.