Trait rivulet::View

pub trait View {
    type Item;
    type Error: Debug;
    fn view(&self) -> &[Self::Item];

    fn poll_grant(
        self: Pin<&mut Self>, 
        cx: &mut Context<'_>, 
        count: usize
    ) -> Poll<Result<(), Self::Error>>;

    fn release(&mut self, count: usize);

    fn grant(&mut self, count: usize) -> Grant<'_, Self>
    where
        Self: Sized + Unpin,
    { ... }

    fn blocking_grant(&mut self, count: usize) -> Result<(), Self::Error>
    where
        Self: Sized + Unpin,
    { ... }
}

Obtain views into asynchronous contiguous-memory streams.

Associated Types

type Item

The streamed type.

type Error: Debug

The error produced by poll_grant.

Loading content...
Required methods

fn view(&self) -> &[Self::Item]

Obtain the current view of the stream.

This view is obtained by successfully polling poll_grant and advanced by successfully polling poll_release.

If this slice is smaller than last successful grant request, the end of the stream has been reached and no additional values will be provided.

fn poll_grant(
self: Pin<&mut Self>,
cx: &mut Context<'_>,
count: usize
) -> Poll<Result<(), Self::Error>>

Attempt to obtain a view of at least count elements.

If the request exceeds the maximum possible grant (if there is one), an error should be returned.

fn release(&mut self, count: usize)

Attempt to advance past the first count elements in the current view.

Panics

If the request exceeds the current grant, this function should panic.

Loading content...
Provided methods

fn grant(&mut self, count: usize) -> Grant<'_, Self> where
Self: Sized + Unpin,

Create a future that obtains a view of at least count elements.

See poll_grant.

fn blocking_grant(&mut self, count: usize) -> Result<(), Self::Error> where
Self: Sized + Unpin,

Obtains a view of at least count elements, blocking the current thread.

See poll_grant.

Loading content...